Adds pyupgrade (#61)
* Adds pyupgrade

pyupgrade is a tool to migrate Python codebases to use the syntactic sugars introduced in the recent Python versions.

https://github.com/asottile/pyupgrade

The current implementation provides both a formatter and a linter. It uses the configured min Python version if available and fallbacks to Python 3+ if not.

There is also an option to disable migration to new typing syntax ("| None", "list[T]"...). This option is currently applied on this repository.

* pyupgrade: Uses dict.pop() to simplify the implementation

* Add release note